# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PrintFlow Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PrintJob {
  id: number;
  createdAt: string;
  printDate: string;
  printTime: string;
  printerName: string;
  pages: number;
  printType: string;
  paperSize: string;
  paperType: string;
  rate: number;
  amount: number;
}

export interface PrintJobList {
  jobs: PrintJob[];
  total: number;
}

export type PrintJobInputPrintType = typeof PrintJobInputPrintType[keyof typeof PrintJobInputPrintType];


export const PrintJobInputPrintType = {
  BW: 'BW',
  Color: 'Color',
  Glossy: 'Glossy',
  Photo: 'Photo',
} as const;

export type PrintJobInputPaperSize = typeof PrintJobInputPaperSize[keyof typeof PrintJobInputPaperSize];


export const PrintJobInputPaperSize = {
  A4: 'A4',
  A3: 'A3',
} as const;

export type PrintJobInputPaperType = typeof PrintJobInputPaperType[keyof typeof PrintJobInputPaperType];


export const PrintJobInputPaperType = {
  Plain: 'Plain',
  Glossy: 'Glossy',
  Photo: 'Photo',
} as const;

export interface PrintJobInput {
  printerName: string;
  pages: number;
  printType: PrintJobInputPrintType;
  paperSize: PrintJobInputPaperSize;
  paperType: PrintJobInputPaperType;
}

export interface DeleteResult {
  success: boolean;
}

export interface Settings {
  id: number;
  bwA4Price: number;
  bwA3Price: number;
  colorA4Price: number;
  colorA3Price: number;
  glossyPrice: number;
  photoPrice: number;
  autoSyncEnabled: boolean;
  syncInterval: number;
}

export interface SettingsInput {
  bwA4Price?: number;
  bwA3Price?: number;
  colorA4Price?: number;
  colorA3Price?: number;
  glossyPrice?: number;
  photoPrice?: number;
  autoSyncEnabled?: boolean;
  syncInterval?: number;
}

export interface DashboardStats {
  totalRevenue: number;
  totalPages: number;
  totalJobs: number;
  glossyRevenue: number;
  revenueGrowth: number;
  pagesGrowth: number;
  jobsGrowth: number;
  glossyGrowth: number;
}

export interface DailySummary {
  date: string;
  bwPages: number;
  colorPages: number;
  glossyPages: number;
  photoPages: number;
  totalPages: number;
  bwAmount: number;
  colorAmount: number;
  glossyAmount: number;
  photoAmount: number;
  totalAmount: number;
}

export interface PrintDistribution {
  bwPercent: number;
  colorPercent: number;
  glossyPercent: number;
  photoPercent: number;
  bwPages: number;
  colorPages: number;
  glossyPages: number;
  photoPages: number;
}

export interface RevenueDataPoint {
  label: string;
  bwAmount: number;
  colorAmount: number;
  glossyAmount: number;
  photoAmount: number;
  totalAmount: number;
}

export interface RevenueAnalytics {
  view: string;
  totalRevenue: number;
  totalPages: number;
  data: RevenueDataPoint[];
}

export interface PrinterAnalytics {
  printerName: string;
  totalPages: number;
  totalJobs: number;
  totalRevenue: number;
  bwPages: number;
  colorPages: number;
  glossyPages: number;
  photoPages: number;
}

export interface PrinterAnalyticsList {
  printers: PrinterAnalytics[];
}

export interface Report {
  filter: string;
  dateFrom: string;
  dateTo: string;
  totalRevenue: number;
  totalPages: number;
  bwPages: number;
  colorPages: number;
  glossyPages: number;
  photoPages: number;
  bwRevenue: number;
  colorRevenue: number;
  glossyRevenue: number;
  photoRevenue: number;
  topPrinter: string;
  avgRevenuePerPrint: number;
  highestRevenueDay: string;
  jobs: PrintJob[];
}

export interface SyncStatus {
  connected: boolean;
  autoSyncEnabled: boolean;
  /** @nullable */
  lastSyncTime: string | null;
  recordsSynced: number;
  syncInterval: number;
}

export interface SyncTriggerInput {
  /** @nullable */
  sheetId?: string | null;
}

export interface SyncResult {
  success: boolean;
  recordsSynced: number;
  syncTime: string;
  message: string;
}

export type ListPrintJobsParams = {
limit?: number;
offset?: number;
dateFrom?: string;
dateTo?: string;
printType?: string;
paperType?: string;
};

export type GetDashboardStatsParams = {
period?: GetDashboardStatsPeriod;
};

export type GetDashboardStatsPeriod = typeof GetDashboardStatsPeriod[keyof typeof GetDashboardStatsPeriod];


export const GetDashboardStatsPeriod = {
  today: 'today',
  week: 'week',
  month: 'month',
  year: 'year',
} as const;

export type GetPrintDistributionParams = {
period?: GetPrintDistributionPeriod;
};

export type GetPrintDistributionPeriod = typeof GetPrintDistributionPeriod[keyof typeof GetPrintDistributionPeriod];


export const GetPrintDistributionPeriod = {
  today: 'today',
  week: 'week',
  month: 'month',
  year: 'year',
} as const;

export type GetRevenueAnalyticsParams = {
view?: GetRevenueAnalyticsView;
year?: number;
month?: number;
};

export type GetRevenueAnalyticsView = typeof GetRevenueAnalyticsView[keyof typeof GetRevenueAnalyticsView];


export const GetRevenueAnalyticsView = {
  weekly: 'weekly',
  monthly: 'monthly',
  yearly: 'yearly',
} as const;

export type GetReportParams = {
filter?: GetReportFilter;
dateFrom?: string;
dateTo?: string;
};

export type GetReportFilter = typeof GetReportFilter[keyof typeof GetReportFilter];


export const GetReportFilter = {
  today: 'today',
  yesterday: 'yesterday',
  week: 'week',
  month: 'month',
  year: 'year',
  custom: 'custom',
} as const;