# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PrintFlow Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /print-jobs:
    get:
      operationId: listPrintJobs
      tags: [print-jobs]
      summary: List print jobs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: printType
          in: query
          schema:
            type: string
        - name: paperType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of print jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintJobList"
    post:
      operationId: createPrintJob
      tags: [print-jobs]
      summary: Add a print job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrintJobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintJob"
  /print-jobs/{id}:
    delete:
      operationId: deletePrintJob
      tags: [print-jobs]
      summary: Delete a print job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
components:
  schemas:
...
    PrintJob:
      type: object
      required: [id, createdAt, printDate, printTime, printerName, pages, printType, paperSize, paperType, rate, amount]
      properties:
        id:
          type: integer
        createdAt:
          type: string
        printDate:
          type: string
        printTime:
          type: string
        printerName:
          type: string
        pages:
          type: integer
        printType:
          type: string
        paperSize:
          type: string
        paperType:
          type: string
        rate:
          type: number
        amount:
          type: number
    PrintJobList:
      type: object
      required: [jobs, total]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/PrintJob"
        total:
          type: integer
    PrintJobInput:
      type: object
      required: [printerName, pages, printType, paperSize, paperType]
      properties:
        printerName:
          type: string
        pages:
          type: integer
        printType:
          type: string
          enum: [BW, Color, Glossy, Photo]
        paperSize:
          type: string
          enum: [A4, A3]
        paperType:
          type: string
          enum: [Plain, Glossy, Photo]
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...