# lib/api-zod/src/generated/types/printJob.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PrintFlow Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface PrintJob {
  id: number;
  createdAt: string;
  printDate: string;
  printTime: string;
  printerName: string;
  pages: number;
  printType: string;
  paperSize: string;
  paperType: string;
  rate: number;
  amount: number;
}