# lib/api-zod/src/generated/types/printJobInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PrintFlow Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { PrintJobInputPaperSize } from './printJobInputPaperSize';
import type { PrintJobInputPaperType } from './printJobInputPaperType';
import type { PrintJobInputPrintType } from './printJobInputPrintType';

export interface PrintJobInput {
  printerName: string;
  pages: number;
  printType: PrintJobInputPrintType;
  paperSize: PrintJobInputPaperSize;
  paperType: PrintJobInputPaperType;
}