# lib/api-zod/src/generated/types/printJobList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PrintFlow Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { PrintJob } from './printJob';

export interface PrintJobList {
  jobs: PrintJob[];
  total: number;
}